pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ppt-skills"
version = "0.1.0"
description = "JSON 設定からビジネス向け PowerPoint を自動生成するライブラリ"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
dependencies = [
    "python-pptx>=0.6.23",
    "jsonschema>=4.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0"]
thumbnail = ["Pillow>=10.0"]

[project.scripts]
ppt-skills = "src.cli:main"

[tool.setuptools]
packages = { find = { include = ["src", "src.*"] } }
package-data = { "src" = ["schema.json"] }
